test: move the proxy env helpers into harness and apply them to the websocket proxy tests
websocket-proxy.test.ts has the same two problems this branch fixes
elsewhere: its NO_PROXY subprocess tests inherit an ambient lowercase
no_proxy that takes precedence over the NO_PROXY they set, and it cleared
NO_PROXY in the module body while restoring it in afterAll, so a run that
loads the file without executing a test in it (bun test -t) leaves the
blanked value behind for later files in the same process. ws-proxy.test.ts
copied that second shape in the previous commit. Both now clear in
beforeAll and restore in afterAll, and the NO_PROXY children start from a
proxy-free env. websocket-syscall-fault.test.ts spawns its CONNECT proxy
client with bunEnv as well, so an ambient NO_PROXY silently turned that
regression test into a direct connection; it gets the same treatment.

clearProxyEnv/restoreProxyEnv/proxyFreeEnv/PROXY_ENV_KEYS move from
proxy-stress-helpers.ts into harness so files outside test/js/bun/http can
use them; the stress helpers re-export them for their existing importers
and proxy.test.ts drops its private copy of the same logic.
